Added precision mode argument named fp

+++ hmsc/updaters/updateZ.py
@@ -66,7 +66,7 @@ def updateZ(params, data, rLHyperparams, *,
     Beta = params["Beta"]
     EtaList = params["Eta"]
     LambdaList = params["Lambda"]
-    sigma = params["sigma"]
+    sigma = tf.cast(params["sigma"], dtype=dtype)
     X = params["Xeff"]
 
     Y = data["Y"]
